test(probe): drop the Sui stalled call test
Every caller bounds the call itself: `network_fingerprints` arms a deadline
before tonic stamps the client's, and the signing flow wraps `extract` in
`FOREIGN_CHAIN_INSPECTION_TIMEOUT`. The client's `DeadlineExceeded` never
surfaces, so the test pinned a tonic detail no caller observes.

A provider that answers `DEADLINE_EXCEEDED` itself does reach the mapping, and
`classified__should_name_a_deadline_as_a_timeout` covers that without a server.
The fake ledger loses the delay it only needed in order to stall.
